media-generator-service/requirements.txt
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
pydantic>=2.8,<3.0
python-docx>=1.1,<2.0
reportlab>=4.2,<5.0
pytest>=8.3,<9.0
uvicorn[standard]>=0.30,<1.0
